feat(analytics): Add Google Analytics
