+++ .gitignore
@@ -1,7 +1,7 @@
 # Specifies intentionally untracked files to ignore when using Git
 # http://git-scm.com/docs/gitignore
-app.firebase.ts
-app.workshop.firebase.ts
+app.config.ts
+app.config.workshop.ts
 
 firebase.json
 .firebaserc
