+++ package.json
@@ -28,6 +28,7 @@
     "@angular/platform-browser-dynamic": "2.2.1",
     "@angular/platform-server": "2.2.1",
     "@ionic/storage": "1.1.7",
+    "angular-ga": "^0.1.3",
     "angularfire2": "^2.0.0-beta.7",
     "firebase": "^3.6.8",
     "ionic-angular": "2.0.0",
