+++ src/app/app.config.sample.ts
@@ -1,0 +1,14 @@
+/**
+ * Rename to `app.config.ts` and fill in the correct information.
+ */
+// Firebase config
+export const firebaseConfig = {
+  apiKey: '',
+  authDomain: '',
+  databaseURL: '',
+  storageBucket: '',
+  messagingSenderId: ''
+};
+
+// Google Analytics Tracking ID
+export const trackingId = '';
