--- src/app/app.firebase.sample.ts
@@ -1,10 +1,0 @@
-/**
- * Rename to `app.firebase.ts` and fill in the correct information.
- */
-export const firebaseConfig = {
-  apiKey: '',
-  authDomain: '',
-  databaseURL: '',
-  storageBucket: '',
-  messagingSenderId: ''
-};
