+++ src/app/app.module.ts
@@ -2,9 +2,10 @@ import { NgModule } from '@angular/core';
 import { IonicApp, IonicModule } from 'ionic-angular';
 import { Storage } from '@ionic/storage';
 import { AngularFireModule } from 'angularfire2';
+import { GoogleAnalyticsModule, GA_TOKEN } from 'angular-ga';
 
 import { ConferenceApp } from './app.component';
-import { firebaseConfig } from './app.firebase';
+import { firebaseConfig, trackingId } from './app.config';
 import { TabsPage } from '../pages/tabs/tabs';
 import { ScheduleModule } from '../pages/schedule';
 import { SessionDetailModule } from '../pages/session-detail';
@@ -21,6 +22,10 @@ import { SharedModule } from '../shared/shared.module';
     IonicModule.forRoot(ConferenceApp),
     AngularFireModule.initializeApp(firebaseConfig),
     SharedModule.forRoot(),
+    GoogleAnalyticsModule.forRoot({
+      provide: GA_TOKEN,
+      useValue: trackingId
+    }),
     ScheduleModule,
     SessionDetailModule,
     SpeakersModule,
