--- src/app/app.workshop.firebase.ts
@@ -1,7 +1,0 @@
-export const firebaseConfig = {
-  apiKey: 'AIzaSyArJjIScro0sU5vtlOg_u6i3sqNV9suKb4',
-  authDomain: 'ng-vikings-workshop.firebaseapp.com',
-  databaseURL: 'https://ng-vikings-workshop.firebaseio.com',
-  storageBucket: 'ng-vikings-workshop.appspot.com',
-  messagingSenderId: '543898875556'
-};
