+++ src/index.html
@@ -50,6 +50,9 @@
   <!-- Ionic's root component and where the app will load -->
   <ion-app></ion-app>
 
+  <!-- Google Analytics Tracking Script -->
+  <script src="assets/libs/ga.js"></script>
+
   <!-- The polyfills js is generated during the build process -->
   <script src="build/polyfills.js"></script>
